requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
reportlab>=4.0
matplotlib>=3.7
schedule>=1.2
python-dateutil>=2.8
